# src/modules/gear/helmets/variants/dto/update-helmet-variant.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateHelmetVariantDto } from './create-helmet-variant.dto';

export class UpdateHelmetVariantDto extends PartialType(CreateHelmetVariantDto) {}