# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/gear/helmets/{id}": {
      "get": {
        "description": "Retorna el detalle completo de un casco con todas sus variantes y tallas.",
        "operationId": "HelmetModelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle del casco"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "summary": "Obtener casco por ID",
        "tags": [
          "Helmets"
        ]
      }
    },
...
    "/api/admin/gear/helmets/{id}": {
      "get": {
        "description": "Incluye todos los campos internos y variantes eliminadas.",
        "operationId": "HelmetModelsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle completo del casco (vista admin)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener casco por ID",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "put": {
        "description": "Todos los campos son opcionales (actualización parcial).",
        "operationId": "HelmetModelsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetModelDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Casco actualizado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar modelo de casco",
        "tags": [
          "Admin — Helmets"
        ]
      },
      "delete": {
        "description": "Marca el casco como eliminado sin borrarlo de la base de datos. Restaurable con /restore.",
        "operationId": "HelmetModelsAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Casco eliminado (soft delete)"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar casco (soft delete)",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/api/admin/gear/helmets/{id}/restore": {
      "post": {
        "description": "Revierte el soft delete, el casco vuelve al estado activo.",
        "operationId": "HelmetModelsAdminController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Casco restaurado"
          },
          "404": {
            "description": "Casco no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Restaurar casco eliminado",
        "tags": [
          "Admin — Helmets"
        ]
      }
    },
    "/api/admin/gear/helmets/{modelId}/variants": {
      "get": {
        "description": "Retorna todas las variantes (colores) del modelo con tallas e inventario.",
        "operationId": "HelmetVariantsAdminController_findAll",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Incluir variantes con soft delete",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de variantes"
          },
          "404": {
            "description": "Modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Listar variantes de un casco",
        "tags": [
          "Admin — Helmet Variants"
        ]
      },
      "post": {
        "description": "Agrega un nuevo color/variante al modelo de casco.",
        "operationId": "HelmetVariantsAdminController_create",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelmetVariantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Variante creada"
          },
          "404": {
            "description": "Modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Crear variante",
        "tags": [
          "Admin — Helmet Variants"
        ]
      }
    },
    "/api/admin/gear/helmets/{modelId}/variants/{variantId}": {
      "get": {
        "operationId": "HelmetVariantsAdminController_findOne",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "UUID de la variante",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la variante con tallas e inventario"
          },
          "404": {
            "description": "Variante o modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Obtener variante por ID",
        "tags": [
          "Admin — Helmet Variants"
        ]
      },
      "put": {
        "description": "Todos los campos son opcionales.",
        "operationId": "HelmetVariantsAdminController_update",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "UUID de la variante",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHelmetVariantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Variante actualizada"
          },
          "404": {
            "description": "Variante o modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Actualizar variante",
        "tags": [
          "Admin — Helmet Variants"
        ]
      },
      "delete": {
        "description": "Restaurable con /:variantId/restore.",
        "operationId": "HelmetVariantsAdminController_remove",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "UUID de la variante",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Variante eliminada (soft delete)"
          },
          "404": {
            "description": "Variante o modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar variante (soft delete)",
        "tags": [
          "Admin — Helmet Variants"
        ]
      }
    },
    "/api/admin/gear/helmets/{modelId}/variants/{variantId}/restore": {
      "post": {
        "operationId": "HelmetVariantsAdminController_restore",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "variantId",
            "required": true,
            "in": "path",
            "description": "UUID de la variante",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Variante restaurada"
          },
          "404": {
            "description": "Variante o modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Restaurar variante eliminada",
        "tags": [
          "Admin — Helmet Variants"
        ]
      }
    },
    "/api/admin/gear/helmets/{modelId}/sizes": {
      "get": {
        "description": "Retorna las tallas disponibles con su inventario detallado por tienda.",
        "operationId": "HelmetSizesAdminController_findAll",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de tallas con inventario"
          },
          "404": {
            "description": "Modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Listar tallas de un modelo",
        "tags": [
          "Admin — Helmet Sizes"
        ]
      },
      "post": {
        "operationId": "HelmetSizesAdminController_create",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHelmetSizeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Talla creada"
          },
          "400": {
            "description": "Talla ya existe para este modelo"
          },
          "404": {
            "description": "Modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Agregar talla a un modelo",
        "tags": [
          "Admin — Helmet Sizes"
        ]
      }
    },
    "/api/admin/gear/helmets/{modelId}/sizes/{sizeId}": {
      "delete": {
        "description": "Elimina permanentemente la talla y su inventario asociado (cascade). No tiene restore.",
        "operationId": "HelmetSizesAdminController_remove",
        "parameters": [
          {
            "name": "modelId",
            "required": true,
            "in": "path",
            "description": "UUID del modelo de casco",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sizeId",
            "required": true,
            "in": "path",
            "description": "UUID de la talla",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Talla eliminada"
          },
          "404": {
            "description": "Talla o modelo no encontrado"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "[Admin] Eliminar talla",
        "tags": [
          "Admin — Helmet Sizes"
        ]
      }
    },
...
  },
  "info": {
    "title": "Moto Gear Picker API",
    "description": "API para selección y gestión de equipamiento de moto",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateHelmetModelDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "brandId": {
            "type": "string"
          },
          "helmetType": {
            "type": "string",
            "enum": [
              "full_face",
              "modular",
              "open_face",
              "half",
              "off_road"
            ]
          },
          "safetyRating": {
            "type": "number",
            "minimum": 1
          },
          "shellMaterial": {
            "type": "string",
            "enum": [
              "polycarbonate",
              "fiberglass",
              "carbon_fiber"
            ]
          },
          "shellSizes": {
            "type": "number",
            "minimum": 1
          },
          "weightGrams": {
            "type": "number",
            "minimum": 0
          },
          "visorAntiScratch": {
            "type": "boolean"
          },
          "visorAntiFog": {
            "type": "boolean"
          },
          "visorPinlock": {
            "type": "string",
            "enum": [
              "pinlock_30",
              "pinlock_30_max_vision",
              "pinlock_70",
              "pinlock_70_max_vision",
              "pinlock_120",
              "pinlock_120_max_vision",
              "pinlock_evo",
              "pinlock_protectint",
              "included_unknown",
              "not_included",
              "not_compatible"
            ]
          },
          "sunVisor": {
            "type": "boolean"
          },
          "sunVisorType": {
            "type": "string"
          },
          "intercomReady": {
            "type": "boolean"
          },
          "intercomDesignedBrand": {
            "type": "string"
          },
          "intercomDesignedModel": {
            "type": "string"
          },
          "removableLining": {
            "type": "boolean"
          },
          "washableLining": {
            "type": "boolean"
          },
          "emergencyRelease": {
            "type": "boolean"
          },
          "closureType": {
            "type": "string",
            "enum": [
              "double_d_ring",
              "micrometric",
              "quick_release",
              "magnetic",
              "pinch_buckle"
            ]
          },
          "certification": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "dot",
                "snell",
                "ece_2206",
                "ece_2205",
                "fim"
              ]
            }
          }
        }
      },
      "CreateHelmetVariantDto": {
        "type": "object",
        "properties": {
          "colorName": {
            "type": "string"
          },
          "colorFamilies": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "black",
                "white",
                "red",
                "blue",
                "green",
                "yellow",
                "orange",
                "pink",
                "purple",
                "grey",
                "silver",
                "gold",
                "brown",
                "carbon",
                "multicolor",
                "bronze",
                "hi_vis"
              ]
            }
          },
          "finish": {
            "type": "string",
            "enum": [
              "gloss",
              "matte",
              "satin"
            ]
          },
          "graphicName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "imageUrl": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "colorName",
          "colorFamilies"
        ]
      },
      "UpdateHelmetVariantDto": {
        "type": "object",
        "properties": {
          "colorName": {
            "type": "string"
          },
          "colorFamilies": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "black",
                "white",
                "red",
                "blue",
                "green",
                "yellow",
                "orange",
                "pink",
                "purple",
                "grey",
                "silver",
                "gold",
                "brown",
                "carbon",
                "multicolor",
                "bronze",
                "hi_vis"
              ]
            }
          },
          "finish": {
            "type": "string",
            "enum": [
              "gloss",
              "matte",
              "satin"
            ]
          },
          "graphicName": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "imageUrl": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateHelmetSizeDto": {
        "type": "object",
        "properties": {
          "sizeLabel": {
            "type": "string",
            "pattern": "/^(XXS|XS|S|M|L|XL|XXL|XXXL|[0-9]{2,3})$/"
          }
        },
        "required": [
          "sizeLabel"
        ]
      },
...
    }
  }
}